update packages and merge header and nav components
