--- src/components/Navbar/Navbar.module.scss
@@ -1,145 +1,0 @@
-@use '../../styles/globals.module' as *;
-
-$duration: 0.2s;
-$distance: 8px;
-$easeOutBack: cubic-bezier(0.175, 0.885, 0.32, 1.275);
-
-.navbar {
-  display: flex;
-  align-items: center;
-  transition: 0.3s ease-in-out;
-  overflow: hidden;
-  font-family: 'PlusJakartaSans';
-}
-
-.navLinks {
-  list-style: none;
-  display: flex;
-  gap: 16px;
-  align-items: baseline;
-  font-size: 1.5rem;
-  line-height: 1.8;
-
-  .link {
-    margin: 0 8px;
-  }
-}
-
-.link {
-  display: inline-block;
-  position: relative;
-  text-decoration: none;
-
-  &:before,
-  &:after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    height: 3px;
-    background-color: $primary;
-  }
-
-  &:before {
-    opacity: 0;
-    transform: translateY(-$distance);
-    transition: transform 0s $easeOutBack, opacity 0s;
-  }
-
-  &:after {
-    opacity: 0;
-    transform: translateY(calc($distance/2));
-    transition: transform $duration $easeOutBack, opacity $duration;
-  }
-
-  &:hover,
-  &:focus {
-    &:before,
-    &:after {
-      opacity: 1;
-      transform: translateY(0);
-    }
-    &:before {
-      transition: transform $duration $easeOutBack, opacity $duration;
-    }
-    &:after {
-      transition: transform 0s $duration $easeOutBack, opacity 0s $duration;
-    }
-  }
-}
-
-.hamburger {
-  display: none;
-  cursor: pointer;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 25px;
-  height: 20px;
-
-  .line {
-    width: 100%;
-    height: 3px;
-    background-color: $white;
-    transition: all 0.3s ease;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .navLinks {
-    position: fixed;
-    top: 0;
-    right: 0;
-    height: 100vh;
-    width: 100%;
-    background-color: $black;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 2rem;
-    transform: translateX(100%);
-    opacity: 0;
-    transition: transform 0.5s ease-out, opacity 0.3s ease-out;
-
-    li {
-      margin: 20px 0;
-
-      a {
-        color: $white;
-        font-size: 2rem;
-      }
-    }
-  }
-
-  .link {
-    border-bottom: 1px solid $primary;
-    padding: 0 2rem 2rem;
-
-    &:last-child {
-      border-bottom: none;
-    }
-  }
-
-  .hamburger {
-    display: flex;
-  }
-}
-
-.active {
-  transform: translateX(0);
-  opacity: 1;
-}
-
-.change {
-  &.line:nth-child(1) {
-    transform: rotate(-45deg) translate(-5px, 6px);
-  }
-
-  &.line:nth-child(2) {
-    opacity: 0;
-  }
-
-  &.line:nth-child(3) {
-    transform: rotate(45deg) translate(-5px, -6px);
-  }
-}
