+++ src/lib/utils/formatDate.ts
@@ -1,17 +1,17 @@
-type optionsType = {
-  year: 'numeric' | '2-digit' | undefined;
-  month: 'numeric' | '2-digit' | 'long' | 'short' | 'narrow' | undefined;
-  day: 'numeric' | '2-digit' | undefined;
-};
+// type optionsType = {
+//   year: 'numeric' | '2-digit' | undefined;
+//   month: 'numeric' | '2-digit' | 'long' | 'short' | 'narrow' | undefined;
+//   day: 'numeric' | '2-digit' | undefined;
+// };
 
-const formatDate = (date: string) => {
-  const options: optionsType = {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  };
+// const formatDate = (date: string) => {
+//   const options: optionsType = {
+//     year: 'numeric',
+//     month: 'short',
+//     day: 'numeric',
+//   };
 
-  return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
-};
+//   return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
+// };
 
-export { formatDate };
+// export { formatDate };
