+++ src/lib/utils/kebabCase.ts
@@ -1,6 +1,6 @@
-// import { slug } from 'github-slugger';
+// // import { slug } from 'github-slugger';
 
-// const kebabCase = (str: string) => slug(str);
-const kebabCase = (str: string) => str;
+// // const kebabCase = (str: string) => slug(str);
+// const kebabCase = (str: string) => str;
 
-export default kebabCase;
+// export default kebabCase;
